[inkling] Rename scripts/run_inkling_975b.py to run_inkling.py
The launcher drives the whole family (Inkling 975B, Inkling-Small,
and the 4-layer slice), so the size suffix no longer fits. Updates
the CI-test imports and the docs command lines; the fully-async
variant keeps its own filename.
