+++ tests/e2e/megatron/model_scripts/test_inkling_4layer_lora_ci.py
@@ -1,12 +1,12 @@
 import os
 
-from scripts.run_inkling_975b import _MODEL_REGISTRY, ScriptArgs, _train
+from scripts.run_inkling import _MODEL_REGISTRY, ScriptArgs, _train
 from tests.ci.ci_register import register_cuda_ci
 from tests.ci.metric_history import register_ci_gate
 
 import miles.utils.external_utils.command_utils as U
 
-# Smoke test for scripts/run_inkling_975b.py --train-mode lora on the 4-layer slice:
+# Smoke test for scripts/run_inkling.py --train-mode lora on the 4-layer slice:
 # shared-outer grouped-expert LoRA served through SGLang's virtual-experts path, one
 # 4-GPU engine, adapter sync verified by checksum. Functionality, not accuracy.
